add the ScanEnumerator and tests, move workflow component objects to utils
