+++ openff/qcsubmit/workflow_components/base_component.py
@@ -9,7 +9,7 @@ from qcelemental.util import which_import
 from typing_extensions import Literal
 
 from openff.qcsubmit.common_structures import ComponentProperties
-from openff.qcsubmit.datasets import ComponentResult
+from openff.qcsubmit.workflow_components.utils import ComponentResult
 
 
 class CustomWorkflowComponent(BaseModel, abc.ABC):
